trying out using doubles for modular exponentiation

+++ RSA/main.c
@@ -8,7 +8,7 @@ int main (int argc, char * argv[])
 	struct privateInfo PI;
 	char input[] = "hello";
 	int messagelength = sizeof(input)/sizeof(char);
-	int cipher[messagelength];
+	long long cipher[messagelength];
 	char message[messagelength];
 	int a = atoi(argv[1]);
 	int b = atoi(argv[2]);
